+++ rosplane_sim/include/rosplane_sim/aircraft_forces_and_moments.h
@@ -34,7 +34,7 @@
 #include <std_msgs/Float32.h>
 #include <geometry_msgs/Vector3.h>
 
-#include "rosplane_sim/gazebo_compat.h"
+#include "rosplane_sim/gz_compat.h"
 
 namespace gazebo
 {
