+++ rosplane_sim/include/rosplane_sim/aircraft_truth.h
@@ -35,7 +35,7 @@
 #include <geometry_msgs/Vector3.h>
 
 #include "rosplane_sim/common.h"
-#include "rosplane_sim/gazebo_compat.h"
+#include "rosplane_sim/gz_compat.h"
 
 namespace gazebo
 {
